+++ changelog.md
@@ -6,6 +6,8 @@
   - JavaScript
   - WebAssembly
   - MicroHs
+- add diffTimeOfDay
+- add \[add/diff\]\[UTC/Local\]Duration\[Clip/Rollover\]
 - UNIXish formatting/parsing:
   - add instance ParseTime DayOfWeek
   - make use of %s specifiers in parsing various types